Created class-based handwriting/handwriting.canvas class through adaptation of functional-programming code by gabor-kovacs. Included canvas element in Drawing.tsx react component so that drawing is possible on a 220px by 220px area.
