+++ src/lib/handwriting-options.ts
